+++ src/controllers/Admin-Web-App/appQueries/api-keys.js
@@ -1,25 +1,6 @@
 const { queryAllUserProjects } = require('../../../helpers/common-queries/project-queries')
+const { formatApiKeyData } = require('./data-cleaners/cleaner-api-keys')
 
-function formatApiKeyData (projects) {
-    let cleanedData = {}
-    cleanedData['numProjects'] = projects.length
-    cleanedData['names'] = []
-    for (let project of projects){
-        let {
-            name, 
-            developmentApiKey, 
-            productionApiKey, 
-        } = project
-        let newPayload = {
-            name, 
-            developmentApiKey, 
-            productionApiKey
-        }
-        cleanedData[name] = newPayload
-        cleanedData['names'].push(name)
-    }
-    return cleanedData
-}
 
 async function getApiKeys (req, res) {
     try {
