+++ src/controllers/Admin-Web-App/appQueries/data-cleaners/cleaner-api-keys.js
@@ -1,0 +1,24 @@
+function formatApiKeyData (projects) {
+    let cleanedData = {}
+    cleanedData['numProjects'] = projects.length
+    cleanedData['names'] = []
+    for (let project of projects){
+        let {
+            name, 
+            developmentApiKey, 
+            productionApiKey, 
+        } = project
+        let newPayload = {
+            name, 
+            developmentApiKey, 
+            productionApiKey
+        }
+        cleanedData[name] = newPayload
+        cleanedData['names'].push(name)
+    }
+    return cleanedData
+}
+
+module.exports = {
+    formatApiKeyData
+}
