+++ src/controllers/Admin-Web-App/appQueries/features.js
@@ -1,43 +1,9 @@
 const { QueryFeaturesByUser } = require('../../../helpers/common-queries/feature-queries')
 const { QueryFeaturesByProject } = require('../../../helpers/common-queries/feature-queries')
-const prettyFormatDate = require('../../../helpers/app-queries-helpers/prettyFormaters')
-
-function removeSensitiveFeatureData (features) {
-    let cleaned_data = {}
-    cleaned_data['numFeatures'] = features.length
-    cleaned_data['names'] = []
-    cleaned_data['allFeatureNames'] = []
-    for (let feature of features){
-        let {
-            name, 
-            variables, 
-            productionEnabled, 
-            developmentEnabled, 
-            parentProjectName, 
-            createdAt, 
-        } = feature
-        let uniqueEntryName = `${name}_${parentProjectName}`
-        cleaned_data[uniqueEntryName] = {
-            name, 
-            variables : variables.length, 
-            productionEnabled, 
-            developmentEnabled, 
-            parentProjectName,
-            createdAt : prettyFormatDate(createdAt), 
-        }
-        cleaned_data['names'].push(uniqueEntryName)
-        cleaned_data['allFeatureNames'].push(name)
-    }  
-    return cleaned_data 
-}
-
-function returnFeatureNames (features) {
-    let names = []
-    for (let feature of features) {
-        names.push(feature.name)
-    }
-    return { names }
-}
+const { 
+    returnFeatureNames, 
+    removeSensitiveFeatureData, 
+} = require('./data-cleaners/cleaner-features')
 
 async function getUserFeatures (req, res) {
     try {
@@ -51,7 +17,6 @@ async function getUserFeatures (req, res) {
     }
 }
 
-
 async function GetFeaturesByProjectName (req, res) {
     try {
         let { user } = req 
@@ -64,7 +29,6 @@ async function GetFeaturesByProjectName (req, res) {
         res.sendStatus(500)
     }
 }
-
 
 module.exports = {
     getUserFeatures,
