+++ src/controllers/Admin-Web-App/appQueries/variables.js
@@ -1,35 +1,5 @@
-
 const { QueryVariablesByUser } = require('../../../helpers/common-queries/variable-queries')
-const prettyFormatDate = require('../../../helpers/app-queries-helpers/prettyFormaters')
-
-function removeSensitiveVariableData (variables) {
-    let cleanedData = {}
-    cleanedData['numVariables'] = variables.length
-    cleanedData['names'] = []
-    cleanedData['allVariableNames'] = []
-    for (let variable of variables) {
-        let {
-            name, 
-            parentFeatureName, 
-            developmentEnabled, 
-            productionEnabled, 
-            parentProjectName, 
-            updatedAt 
-        } = variable
-        let uniqueEntryName = `${name}_${parentFeatureName}` 
-        cleanedData[uniqueEntryName] = {
-            name, 
-            parentFeatureName, 
-            developmentEnabled, 
-            productionEnabled, 
-            parentProjectName, 
-            updatedAt : prettyFormatDate(updatedAt) 
-        }
-        cleanedData['names'].push(uniqueEntryName)
-        cleanedData['allVariableNames'].push(name)
-    }
-    return cleanedData
-}
+const { removeSensitiveVariableData } = require('./data-cleaners/cleaner-variables')
 
 async function getVariables (req, res) {
     try {
@@ -42,7 +12,6 @@ async function getVariables (req, res) {
         res.sendStatus(500)
     }
 } 
-
 
 module.exports = {
     getVariables,
